Fix wrong parameter type

+++ visuals.cpp
@@ -306,7 +306,7 @@ int Visuals::main(int argc, char* argv[])
             ("led-controlport,c", po::value<int>(&m_portControl)->default_value(m_portControl), "led control port")
             ("motion-port,g", po::value<int>(&m_portMotion)->default_value(m_portMotion), "motion port")
             ("fps,f", po::value<int>(&m_fps)->default_value(m_fps), "motion port")
-            ("volume,V", po::value<int>(&m_volume)->default_value(m_volume), "initial volume")
+            ("volume,V", po::value<float>(&m_volume)->default_value(m_volume), "initial volume")
         ;
 
         po::options_description podesc_cmdline;
